requirements.txt
langchain
langchain-openai
langchain-pinecone
langchain-community
beautifulsoup4
python-dotenv